:sparkles: criar migrations, models, seeders empresa e entregador

+++ database/migrations/2022_01_11_162000_create_clientes_table.php
@@ -16,7 +16,7 @@ class CreateClientesTable extends Migration
         Schema::create('clientes', function (Blueprint $table) {
             $table->tinyIncrements('id');
             $table->string('nome');
-            $table->integer('empresa_id');
+            $table->tinyInteger('empresa_id');
             $table->foreign('empresa_id')->references('id')->on('empresas');
         });
     }
